+++ go/lists/base/parameter_list.go
@@ -2,21 +2,21 @@ package base
 
 import (
 	"github.com/AssetMantle/schema/go/lists"
-	"github.com/AssetMantle/schema/go/types"
-	"github.com/AssetMantle/schema/go/types/base"
+	"github.com/AssetMantle/schema/go/parameters"
+	"github.com/AssetMantle/schema/go/parameters/base"
 )
 
 var _ lists.ParameterList = (*ParameterList)(nil)
 
-func (parameterList *ParameterList) Get() []types.Parameter {
-	Parameters := make([]types.Parameter, len(parameterList.Parameters))
+func (parameterList *ParameterList) Get() []parameters.Parameter {
+	Parameters := make([]parameters.Parameter, len(parameterList.Parameters))
 	for i, parameter := range parameterList.Parameters {
 		Parameters[i] = parameter
 	}
 	return Parameters
 }
 
-func NewParameterList(parameters ...types.Parameter) lists.ParameterList {
+func NewParameterList(parameters ...parameters.Parameter) lists.ParameterList {
 	newParameters := make([]*base.Parameter, len(parameters))
 	for i, parameter := range parameters {
 		newParameters[i] = parameter.(*base.Parameter)
