+++ go/lists/parameter_list.go
@@ -3,10 +3,10 @@ package lists
 import (
 	"github.com/gogo/protobuf/proto"
 
-	"github.com/AssetMantle/schema/go/types"
+	"github.com/AssetMantle/schema/go/parameters"
 )
 
 type ParameterList interface {
 	proto.Message
-	Get() []types.Parameter
+	Get() []parameters.Parameter
 }
